p3-task4: impl info for displaying system info

+++ include/os/sched.h
@@ -189,7 +189,7 @@ void print_pcb_list(const list_t* queue);
 #ifdef S_CORE
 extern pid_t do_exec(int id, int argc, uint64_t arg0, uint64_t arg1, uint64_t arg2);
 #else
-extern pid_t do_exec(const char* name, int entrance, int argc, char* argv[], unsigned affinity_mask);
+extern pid_t do_exec(const char* name, uint64_t entrance, int argc, char* argv[], unsigned affinity_mask);
 #endif
 extern void do_exit(void);
 extern int do_kill(pid_t pid);
