+++ kernel/sched/sched.c
@@ -281,7 +281,7 @@ void cleanup_proc(pcb_t* pcb) {
     free_pcb(pcb);
 }
 
-pid_t do_exec(const char* name, int entrance, int argc, char* argv[], unsigned affinity_mask) {
+pid_t do_exec(const char* name, uint64_t entrance, int argc, char* argv[], unsigned affinity_mask) {
     pretty_log(LOG_DEBUG, "handling exec for %s", name);
     pcb_t* pcb = construct_pcb(name, entrance, argc, argv, 1, 4);
     if (!pcb) {
